requirements.txt
# Core dependencies
openai>=1.0.0
python-dotenv>=1.0.0
requests>=2.31.0

# Web framework
fastapi>=0.104.0
uvicorn>=0.24.0

# Data processing
pydantic>=2.0.0
numpy>=1.24.0
pandas>=2.0.0

# Utilities
python-multipart>=0.0.6
aiofiles>=23.0.0

# Development dependencies (optional)
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.0.0
pre-commit>=3.3.0 